requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.15.0
streamlit>=1.12.0
pillow>=8.4.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.60.0
requests>=2.25.0